Core/AngelScript: Change opcode handler storage from uint16 to uint32
Changed opcode parameter and storage type from uint16 to uint32 in ASPacketHooks to support full opcode range. Updated RegisterOpcodeHandler(), UnregisterOpcodeHandler(), GetOpcodeHandler() and ShouldBlockOriginal() methods along with OpcodeHandlerEntry struct and _opcodeHandlers map. Removed unnecessary static_cast operations in ASPacketAPI.cpp and AngelScriptMgr.cpp.

+++ AngelScript/API/ASPacketAPI.cpp
@@ -54,11 +54,11 @@ namespace AngelScript
     static void RegisterOpcodeHandler(uint32 opcode, asIScriptFunction* func, bool blockOriginal)
     {
         if (!func) return;
-        ASPacketHooks::instance()->RegisterOpcodeHandler(static_cast<uint16>(opcode), func, blockOriginal);
+        ASPacketHooks::instance()->RegisterOpcodeHandler(opcode, func, blockOriginal);
     }
     static void UnregisterOpcodeHandler(uint32 opcode)
     {
-        ASPacketHooks::instance()->UnregisterOpcodeHandler(static_cast<uint16>(opcode));
+        ASPacketHooks::instance()->UnregisterOpcodeHandler(opcode);
     }
 
     // ---- PacketData wrappers (from ASPacketData.h) ----
